Split up CreateGroupsFromCad and AddRelatedNodesAndElements
