+++ Tools/CreateGroupsFromCAD.cs
@@ -56,56 +56,6 @@
             }
 
             CreateGroupsFromNamedPlanes(femPart);
-
-            AddRelatedNodesAndElements(femPart);
-        }
-
-        /// <summary>
-        /// This function cycles through all cae groups in a CaePart.
-        /// For each group it adds the related nodes and elements for the bodies and faces in the group.
-        /// Practical for repopulating groups after a (partial) remesh.
-        /// Function is idempotent.
-        /// </summary>
-        /// <param name="caePart">The CaePart to perform this operation on.</param>
-        public static void AddRelatedNodesAndElements(CaePart caePart)
-        {
-            CaeGroup[] caeGroups = caePart.CaeGroups.ToArray();
-            foreach (CaeGroup item in caeGroups)
-            {
-                theLW.WriteFullline("Processing group " + item.Name);
-                List<CAEBody> seedsBody = new List<CAEBody>();
-                List<CAEFace> seedsFace = new List<CAEFace>();
-
-                foreach (TaggedObject taggedObject in item.GetEntities())
-                {
-                    if (taggedObject is CAEBody)
-                    {
-                        seedsBody.Add((CAEBody)taggedObject);
-                    }
-                    else if (taggedObject is CAEFace)
-                    {
-                        seedsFace.Add((CAEFace)taggedObject);
-                    }
-                }
-
-                SmartSelectionManager smartSelectionManager = caePart.SmartSelectionMgr;
-
-                RelatedElemMethod relatedElemMethodBody = smartSelectionManager.CreateRelatedElemMethod(seedsBody.ToArray(), false);
-                RelatedNodeMethod relatedNodeMethodBody = smartSelectionManager.CreateRelatedNodeMethod(seedsBody.ToArray(), false);
-                // comment previous line and uncomment next line for NX version 2007 (release 2022.1) and later
-                // RelatedNodeMethod relatedNodeMethodBody = smartSelectionManager.CreateNewRelatedNodeMethodFromBodies(seedsBody.ToArray(), false, false);
-                
-                item.AddEntities(relatedElemMethodBody.GetElements());
-                item.AddEntities(relatedNodeMethodBody.GetNodes());
-
-                RelatedElemMethod relatedElemMethodFace = smartSelectionManager.CreateRelatedElemMethod(seedsFace.ToArray(), false);
-                RelatedNodeMethod relatedNodeMethodFace = smartSelectionManager.CreateRelatedNodeMethod(seedsFace.ToArray(), false);
-                // comment previous line and uncomment next line for NX version 2007 (release 2022.1) and later
-                // RelatedNodeMethod relatedNodeMethodFace = smartSelectionManager.CreateNewRelatedNodeMethodFromFaces(seedsFace.ToArray(), false, false);
-
-                item.AddEntities(relatedElemMethodFace.GetElements());
-                item.AddEntities(relatedNodeMethodFace.GetNodes());
-            }
         }
 
         /// <summary>
